implemented delta and capped framerate at 60fps

+++ src/members/vector2.cpp
@@ -1,7 +1,7 @@
 #include "vector2.hpp"
 
 
-Vector2::Vector2(int x, int y) {
+Vector2::Vector2(float x, float y) {
     this->x = x;
     this->y = y;
 }
