+++ src/members/vector2.hpp
@@ -6,13 +6,13 @@ using namespace std;
 class Vector2 {
 public:
     // Constructor
-    Vector2(int x = 0, int y = 0);
+    Vector2(float x = 0, float y = 0);
 
     // Destructor
     ~Vector2();
 
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 #endif // Vector2_HPP
